backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
paho-mqtt==2.0.0
apscheduler==3.10.4
